+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Catch error when video is unavailable.
+
 ## [2.11.7] - 2026-07-04
 
 ### Fixed